refactor code v1
