+++ tests/unit_tests/test_flux_dataloader.py
@@ -8,13 +8,13 @@
 import torch
 from torchtitan.config_manager import JobConfig
 from torchtitan.datasets.flux_datasets import build_flux_dataloader
-from torchtitan.datasets.tokenizer.HFEmbedder import HFEmbedder
+from torchtitan.experiments.flux.utils import load_clip, load_t5
 
 
 class TestFLUXDataLoader:
     def test_FLUX_dataloader(self):
         dataset_name = "cc12m"
-        batch_size = 2  # batch_size = 4 will cause OOM
+        batch_size = 1  # batch_size = 4 will cause OOM
         seq_len = 512
         world_size = 4
         rank = 0
@@ -31,18 +31,6 @@ class TestFLUXDataLoader:
 
         assert len(input_data) == 6
         # TODO(jianiw): Add more data shape check once we finalize input and model
-
-    def _load_t5(self, device: str = "cpu", max_length: int = 512) -> HFEmbedder:
-        # max length 64, 128, 256 and 512 should work (if your sequence is short enough)
-        return HFEmbedder(
-            "google/t5-v1_1-small", max_length=max_length, torch_dtype=torch.bfloat16
-        ).to(device)
-
-    def _load_clip(self, device: str = "cpu") -> HFEmbedder:
-        # The max length is set to be 77
-        return HFEmbedder(
-            "openai/clip-vit-large-patch14", max_length=77, torch_dtype=torch.bfloat16
-        ).to(device)
 
     def _build_dataloader(
         self, dataset_name, batch_size, seq_len, world_size, rank, seed, device="cpu"
@@ -64,10 +52,8 @@ class TestFLUXDataLoader:
         return build_flux_dataloader(
             dp_world_size=world_size,
             dp_rank=rank,
-            embedder=[
-                self._load_t5(device=device, max_length=seq_len),
-                self._load_clip(device=device),
-            ],
+            t5_encoder=load_t5(device=device, max_length=seq_len),
+            clip_encoder=load_clip(device=device),
             job_config=config,
             infinite=False,
             device=device,
