+++ torchtitan/datasets/flux_datasets.py
@@ -7,24 +7,18 @@
 import math
 import random
 from dataclasses import dataclass
-from logging import raiseExceptions
-from typing import Any, Callable, Dict, List, Optional
+from typing import Any, Callable, Optional
 
 import torch
 from datasets import Dataset, load_dataset
 from datasets.distributed import split_dataset_by_node
 from einops import rearrange, repeat
-from PIL import Image
 from torchdata.stateful_dataloader import StatefulDataLoader
 
 from torchtitan.components.dataloader import BaseDataLoader
 from torchtitan.config_manager import JobConfig
-from torchtitan.datasets.tokenizer.HFEmbedder import HFEmbedder
+from torchtitan.experiments.flux.modules.HFEmbedder import HFEmbedder
 from torchtitan.tools.logging import logger
-from torchtitan.tools.profiling import (
-    maybe_enable_memory_snapshot,
-    maybe_enable_profiling,
-)
 from torchvision import transforms
 
 
@@ -32,15 +26,12 @@ def _process_cc12m_image(
     sample: dict[str, Any], output_size: int = 256
 ) -> tuple[str, torch.Tensor]:
     """
-    Process CC12M dataset sample image.
+    Preprocess CC12M dataset sample image.
     Steps:
         1. Ignore all the images smaller than output_size * output_size.
         2. Resize and crop the image to output_size * output_size.
     """
-    img = sample[
-        "jpg"
-    ]  # NOTE: type is PIL Image, https://huggingface.co/docs/datasets/en/image_load
-
+    img = sample["jpg"]
     prompt = sample["txt"]
 
     if img.width < output_size or img.height < output_size:
@@ -75,7 +66,6 @@ class MultiModalDatasetConfig:
     data_processor: Callable
 
 
-# Add your dataset here here - more information at docs/datasets.md
 DATASETS = {
     "cc12m": MultiModalDatasetConfig(
         path="pixparse/cc12m-wds",
@@ -124,7 +114,8 @@ class FLUXDataLoader(StatefulDataLoader, BaseDataLoader):
         self,
         dataset_name: str,
         dataset_path: Optional[str],
-        embedder: List[HFEmbedder],  # Tokenizer for text
+        t5_encoder: HFEmbedder,
+        clip_encoder: HFEmbedder,
         dp_rank: int = 0,
         dp_world_size: int = 1,
         infinite: bool = False,
@@ -145,10 +136,8 @@ class FLUXDataLoader(StatefulDataLoader, BaseDataLoader):
         self.dataset_name = dataset_name
         self._data = split_dataset_by_node(ds, dp_rank, dp_world_size)
 
-        assert len(embedder) == 2
-
-        self._t5_embedder = embedder[0]
-        self._clip_embedder = embedder[1]
+        self._t5_encoder = t5_encoder
+        self._clip_encoder = clip_encoder
         self._data_processor = data_processor
         self._image_transformer = transforms.ToTensor()
 
@@ -171,12 +160,12 @@ class FLUXDataLoader(StatefulDataLoader, BaseDataLoader):
             next(it)
         return it
 
-    def _prepare(self):
+    def _preprocess(self):
         target_img = torch.stack(self._all_target_imgs).to(self.device)
 
-        sample_t5_embedding = self._t5_embedder(self._all_prompts)
+        sample_t5_embedding = self._t5_encoder(self._all_prompts)
         txt_ids = torch.zeros(self.batch_size, sample_t5_embedding.shape[1], 3)
-        sample_clip_embedding = self._clip_embedder(self._all_prompts)
+        sample_clip_embedding = self._clip_encoder(self._all_prompts)
 
         noise_latent = _get_noise_latent(
             self.batch_size,
@@ -186,11 +175,11 @@ class FLUXDataLoader(StatefulDataLoader, BaseDataLoader):
             seed=self.seed,
         )
 
-        _, _, h, w = noise_latent.shape  # (16, 256/8, 256/8) = (16, 32, 32)
+        _, _, h, w = noise_latent.shape  # (bsz, 16, 256/8, 256/8) = (bsz, 16, 32, 32)
         # patchify the noise latent, p = 2
         noise_latent = rearrange(
             noise_latent, "b c (h ph) (w pw) -> b (h w) (c ph pw)", ph=2, pw=2
-        )
+        )  # (bsz, 16, 32, 32) -> (bsz, 16, 16, 64)
 
         img_ids = torch.zeros(h // 2, w // 2, 3)
         img_ids[..., 1] = img_ids[..., 1] + torch.arange(h // 2)[:, None]
@@ -226,7 +215,7 @@ class FLUXDataLoader(StatefulDataLoader, BaseDataLoader):
                 self._all_prompts.extend(sample_prompt)
 
                 if samples_cnt == self.batch_size:
-                    batched_data = self._prepare()
+                    batched_data = self._preprocess()
                     samples_cnt = 0
                     self._all_prompts = []
                     self._all_target_imgs = []
@@ -256,7 +245,8 @@ class FLUXDataLoader(StatefulDataLoader, BaseDataLoader):
 def build_flux_dataloader(
     dp_world_size: int,
     dp_rank: int,
-    embedder: List[HFEmbedder],
+    t5_encoder: HFEmbedder,
+    clip_encoder: HFEmbedder,
     job_config: JobConfig,
     infinite: bool = True,
     device: str = "cpu",
@@ -270,7 +260,8 @@ def build_flux_dataloader(
     return FLUXDataLoader(
         dataset_name=dataset_name,
         dataset_path=dataset_path,
-        embedder=embedder,
+        t5_encoder=t5_encoder,
+        clip_encoder=clip_encoder,
         dp_rank=dp_rank,
         dp_world_size=dp_world_size,
         infinite=infinite,
