+++ torchtitan/experiments/flux/sampling.py
@@ -2,10 +2,8 @@ import math
 from typing import Callable
 
 import torch
-from einops import rearrange
-from torch import Tensor
 
-from torchtitan.experiments.flux.model import Flux
+from torch import Tensor
 
 
 def time_shift(mu: float, sigma: float, t: Tensor):
@@ -37,51 +35,3 @@ def get_schedule(
         timesteps = time_shift(mu, 1.0, timesteps)
 
     return timesteps.tolist()
-
-
-def denoise(
-    model: Flux,
-    # model input
-    img: Tensor,
-    img_ids: Tensor,
-    txt: Tensor,
-    txt_ids: Tensor,
-    vec: Tensor,
-    target: Tensor,  # TODO(jianiw): This is place holder now
-    # sampling parameters
-    timesteps: list[float],
-    guidance: float = 4.0,
-    # extra img tokens
-    img_cond: Tensor | None = None,
-):
-    # this is ignored for schnell
-    guidance_vec = torch.full(
-        (img.shape[0],), guidance, device=img.device, dtype=img.dtype
-    )
-    for t_curr, t_prev in zip(timesteps[:-1], timesteps[1:]):
-        t_vec = torch.full((img.shape[0],), t_curr, dtype=img.dtype, device=img.device)
-        pred = model(
-            img=torch.cat((img, img_cond), dim=-1) if img_cond is not None else img,
-            img_ids=img_ids,
-            txt=txt,
-            txt_ids=txt_ids,
-            y=vec,
-            timesteps=t_vec,
-            guidance=guidance_vec,
-        )
-
-        img = img + (t_prev - t_curr) * pred
-
-    return img
-
-
-def unpack(x: Tensor, height: int, width: int) -> Tensor:
-    return rearrange(
-        x,
-        "b (h w) (c ph pw) -> b c (h ph) (w pw)",
-        h=math.ceil(height / 16),
-        w=math.ceil(width / 16),
-        ph=2,
-        pw=2,
-    )
-    # Input tensor shape: torch.Size([224, 16, 64]). Additional info: {'h': 32, 'w': 32, 'ph': 2, 'pw': 2}.
